[Decomp][Unit] Cohesion split: extract hostility/reaction methods to UnitHostility.cpp
